main/requirements.txt
python-telegram-bot==20.7
playwright==1.40.0
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3
